Added BusyELKS to the repository. BusyELKS is an attempt to save space by generating a series of BusyBox-like binaries with symbolic links to replace stand-alone binaries. BusyELKS was generated by combining most of the "elkscmd" program files together into one set, eliminating redundant chunks of code from where each program was split off from a larger code base (i.e. sash was used to provide various utilities and code was duplicated to make this happen), and writing a build script and some "glue" code to handle the detection of what command is being requested.
BusyELKS has not been tested, though it compiles as-is and seems to
function under elksemu on Linux-i386, currently the /sbin/init
process is known to not work properly. Attempts to trace the reason
for the failure point to a possible bug in the ELKS kernel, but
if init is left as a standalone program, the rest of BusyELKS very
well may work out-of-the-box.

The space savings from combining separate programs into one binary
is quite significant; in fact, BusyELKS and the ELKS kernel combined
can likely fit on one 360K floppy.

Signed-off-by: Jody Bruchon <jody@jodybruchon.com>
Committed-by: Jody Bruchon <jody@jodybruchon.com>