Fix to exec syscall, improve find_buffer performance
In function sys_execve() in file fs/exec.c, if after opening an
executable file an error was found, closed the file and returned an error
code, but the corresponding inode structure was not released. Now,
the inode is released on errors.

In file fs/buffer.c, function find_buffer() now looks for the
requested block in the direction most likely to find it faster.

Moved prototype for function get_unused_fd() from include/arch/system.h
to the more reasonable place in include/linuxmt/fs.h.

Code size was reduced by 16 bytes.