rootfs_template: remove some unnecessary stuff, fix root shell
ELKS has two shells: ash and sash. The proper system shell should
be symlinked at /bin/sh and the default shell for root should be
set to this symlink. Some totally unnecessary bogus users exist in
the passwd/group files and just happened to make those files a bit
larger than a 1K disk block, so a few of these were removed to
lower the default size of these files and possibly save a tiny bit
of disk space.