Change 1.722 floppy to 1.680, add comb360 root+boot image type
1.722MB floppies use a geometry of 82 tracks and 21 sectors which
makes it less likely to work in some older machines with the MINIX
boot sector code we use from dev86. Switch to 1.680MB format which
uses 80 tracks so that the track count matches every other floppy
format.

Add a bootable 360K floppy with a root filesystem on it. The new
"comb360" image type is a fully self-contained bootable 360K disk
with a very minimal set of programs. ELKS can finally boot on a
single 360K floppy disk.