Fix problem with sys_brk() failing unnecessarily, add images_clean
A check was being done in sys_brk() that involved subtracting
0x1000 from t_endseg before checking against it. There is no
information available regarding why this was done this way and it
is preventing some programs from running correctly; specifically,
a "login:" loop was being caused by this check failing. Removing
the unexplained offset makes programs run correctly and nothing
seems to be misbehaving as a result. More of these strange issues
from code written over 10 years ago are likely to appear in the
future, unfortunately.

Added a 'make images_clean' in elkscmd to delete any built images.
Especially handy for rebuilding images with a new kernel but no
changes to other code.