Preparation for named pipes
In file fs/inode.c, function read_inode() dereferenced a pointer and
later checked that the pointer was not NULL. Now, it first makes the
check.

Remove a very small usage of members i_pipe and i_sock in inode
structures, as well as those members in preparation for the
implementation of named pipes and sockets.

In file fs/minix/namei.c, in many functions, release resources in the
inverse order as they were allocated.

In fs/pipe.c, pipe buffers are now allocated in the pipe_open
function, instead of inode creation function get_pipe_inode().

In init/main.c, after failing to find the init program, opened
/dev/tty0 as console and starts /bin/sh, but in ELKS, the console is
in /dev/tty1, not /dev/tty0. Also, fixed the order of search of the
init program.

Code size was reduced by 32 bytes and bss reduced by 192 bytes.