Improve fdisk behavior and text, fix init stack overflow
Somehow, init's stack overflow was un-fixed, so it has been fixed
again in this commit.

fdisk suffered from poor text formatting and lots of unnecessary
code. Its commands did not match fdisk from util-linux and BusyBox
which is frustrating. An unlikely condition where a geometry ioctl
returns a zero for any of the CHS values was allowing the program
to continue anyway when it should immediately abort with an error.
The code style is compressed and ugly, so some (but not all) code
was cleaned up for better readability. The "die" #define and some
exit points were replaced with 'goto usage' and a common usage
code section was added.