Fix compile-time generation of system call table
The syscall table 'arch/i86/kernel/entry.c' is supposed to be
generated at compile time but was not being generated correctly.
This change causes the syscall table to be cleaned during a
'make clean' and properly rebuilt when the file is missing. This
also fixes errors observed after the patch "Changes to reduce
warnings using ia16-unknown-elks-gcc" was applied.