Final changes to compile ELKS kernel with ia16-gcc
-Rename assembly files named *.c to *.S
-Adjust Makefiles to account for this changes of name

The kernel was compiled with BCC and tested with QEMU.
Code size unchanged. The ia16-gcc compiler successfully
builds the kernel, with a code size reduction of 3056
bytes relative to the kernel produced with bcc.
Under QEMU, the ia16-gcc produced kernel completes the
initialization, but fails to mount the root.