elkscmd: Add APM shutdown to sys_reboot() and a 'poweroff' utility
When testing ELKS in an emulator, it has become annoying that
there is no way to "shut down" the system and power off using
"modern" power management mechanisms. This commit adds a flag code
0xDEAD to the sys_reboot() call which invokes apm_shutdown_now()
which performs three Advanced Power Management (APM) BIOS calls to
power off the computer. Also included is the 'poweroff' command
which performs this system call. Tested in QEMU and successfully
found to "power off" the emulator and exit gracefully.