Fix full5 build, clean up HD image script, fix httpd Makefile
This fixes a failure in the full5 build by dropping byacc and m4
from the default "standard" programs. It also fixes a bug in the
httpd Makefile where a directory didn't exist before a file was
copied to it which resulted in an error.

.gitignore has been updated to be silent about the hard drive
image file name.

This also includes improvements to Georg's hard drive image build
script to improve error checking and flexibility for any future
changes that may be needed.