Add a simple master Makefile and silence an unneeded warning
Developers will often try to 'make' and 'make clean' instead of
running the build.sh script provided. This adds a Makefile that
simply wraps around the build script to soothe the pain a bit.

This also silences the "must be root" warning for building disk
images when a clean is requested because it's not relevant there.