Update Makefiles to fix 360K image builds
The PTY shell test shouldn't be included on minimal installs like
the 360K floppy images, so remove it. This commit also removes the
'l' command from minimal images since it is not a standard command
and is essentially a stripped 'ls' with a different output format.
There is no obvious benefit to it and it uses up several KiB on
the disk.