Replace select() queues hash by table
Result of the hash function used in select() was the same for two queues
within the same 256 bytes address range, and was unsafe for other queues
scattered over the kernel data segment.

The hash is replaced by a table of the polled queues. This design is the
safest, at the cost of some extra bytes in the task structure (+8).

Fixes #208