[kernel] syscall.dat now indicates variadic syscall routines
This affects the generation of the code for e.g. the open
(...) function.  Although open (path, flags) and open (path,
flags, mode) can both be called the same way as, say,
write (fd, buf, count) under the `cdecl` calling convention,
this is not the case under the other calling conventions
(`rtd`, `regparmcall`).