[boot] put boot_minix.c's entry pt. at sector start
This means that the OS-independent boot code (boot_sect.S)
does not need to know where precisely the OS-specific
load_prog() routine is.  It can instead simply do a `call'
to the start of the second sector.

This means we need to make sure that gcc-ia16 places the
`jmp load_prog' right at the start.  (Currently I use
-fno-toplevel-reorder, which also suppresses the reordering
of all other functions in boot_minix.c...)