[elks] remove CONFIG_SHORT_FILES checks in defining fd_set etc.
Documentation/historic/CHANGELOG says of ELKS 0.0.79, "Added
CONFIG_SHORT_FILES which limits file descriptors to 16."
However, this is no longer a configuration option now, so
there is no real need to check for it.

Removing this check simplifies the code for processing
fd_set's in select(...) calls, and removes a potential source
of ABI instability.