env.sh: Do not error out if CROSSDIR is missing
This error results in build.sh not being able to make CROSSDIR
for the user and continue the easy build process. The user should
not need to make CROSSDIR themselves, or at least such a check
belongs in other scripts, not in the env.sh script that is called
to set up the environment.