[boot] [elks] allow bootloaders to load /linux as one blob
Previously, bootloaders have to do the job of splitting up
/linux into its "setup" portion (which goes to INITSEG:0 and
SETUPSEG:0) and its "kernel" portion (SYSSEG:0) in memory.

With this change, a bootloader can load /linux as one single
binary blob at INITSEG:0, and inform /linux's "setup" portion
about this (via a new elks_flags field in the loaded /linux).

This helps simplify the logic (and shrink the sizes) of the
Minix and FAT filesystems' boot code.