[boot] modify _and_ restore floppy parameter table
This is a follow-up to the earlier quick bugfix in the
commit 65da1a7d06567bb138f73abda441b4acd9dea1ef.

This new commit modifies the Diskette Drive Parameter Table
(DDPT, int $0x1e) to give the correct number of sectors per
track, and uses the new DDPT to read /linux.

It also now _restores_ the old DDPT just before handing over
to /linux, or (upon an error) before rebooting.

To accommodate the code to restore the old DDPT, I also had
to trim down the disk read logic for the FAT bootloader ---
it now reads /linux more slowly, one sector at a time.