[kernel] [libc] [elksemu] (ABI CHANGE) Make _signal syscall accept a far pointer to a far signal handler
This is in preparation for adding support for "medium"
memory model (far code + near data) programs to ELKS
(see https://github.com/tkchia/build-ia16/issues/14).

Currently the kernel assumes that, when a user process
handles an asynchronous signal, the signal handler can
simply return to a near (16-bit) address.  But this may
not be true once ELKS allows a user process to run code in
a far text segment.

This patch changes the semantics of the _signal syscall so
that it accepts a far (32-bit) signal handler pointer which
returns to a far address.

To save data space, I also changed the kernel's internal
signal handling structures.