[boot] FAT12/16 boot: search 1st 8 root dir. entries for /linux
(Reworked from https://github.com/jbruchon/elks/pull/854 .)

With this patch, the bootloader will be able to find /linux
as long as it is in the first 8 root directory entries of a
FAT12 or FAT16 filesystem (including hidden entries for long
file names etc.).  However, /linux's contents must still be
laid out in contiguous sectors.

The bootloader now also checks that the /linux directory
entry it finds is for a normal file.

This patch probably needs more stress-testing.