ssd: add generic SD card ssd driver. 8018x: add bit-banged low level SD card driver for ssd.
Adds support for generic SD cards as ssd block devices.
Adds a low-level SPI bit-banged driver that allows the usage
of the generic SD card driver on the 8018x platform. GPIO
lines are configurable on the source code to some degree.
Confirmed to work with a 2GB SanDisk SD card, formatted
with fat16.