[elksemu] fix: checks on elks_base both too strict & too loose
On Linux/i386 --- and on Linux/x86-64 with gcc -m32 --- do
not abort elksemu if the ELKS code and data would be mapped
to an address with the high bit (bit 31) set.  Such addresses
might be perfectly good user space addresses.

On Linux/x86-64, check that the entire ELKS memory does not
cross the 4 GiB mark, rather than just checking the start of
ELKS memory.