Fix tr so that squeeze does not act like delete.
When there is only one set, the test len1 > len2 will usually be true, so set2
would get changed from NULL to a padded copy of set1, but since set2 was empty
anyway it becomes completely '\0' filled.  Since there is now a non-NULL set2,
the "translate" clause of the main loop runs, which would translate the read
character to '\0', which is also used to mean "do not print".