[libc] Fix fd leak: Close even if flush fails
POSIX says that if fclose fails, the state of the handle is undefined,
so the app is not allowed to try again to close it.  And it's widely
known that if you really care about data integrity, you must manually
flush and error check first, before closing.  So no matter what, fclose
should close.