[8018x] add romprg to flash memory chips thru serial
This adds support for a port of the "romprg" project, which enables
ELKS to be used as an in-system flash programmer for the same memory
that was used to run it. It also provides a PC-side tool that is
the "command line tool" for the "romprg" running on the ELKS system.