[8018x] enable AT-compatible RTC at specific IO address
This makes the clock application read the proper registers of the
RTC directly via its I/O address; without passing thru the CMOS
I/O mechanism found in AT machines. The RTC needs to be an
AT-compatible just for the sake of the register map used.
Also updates the 8018x.config, which sets the proper IO address for
the RTC on my SBC, and enables the medium model exec compatibility.