Fix installation (or not) of tests; saves 52k on 1440 or 104k when all disabled
Reorganize the tests: float looks like dev testing code, not a user
tool, but was unconditionally installed to the 1440k image.  Install
some of the tests with the :test tag.

Aside: CONFIG_APPS_<SIZE> does not set the various CONFIG_APP_* so tests
were not added to INSTDIRS, so tests do not 'make install' regardless of
CONFIG_APP_TESTS.  Confusing yet ultimately not the root cause.