Fixed the polling lp driver.
The lp driver was non-functional primarily due to not checking the busy
status before writing a character.  The old code was only checking for
errors, not busy, and then blasting ahead and dropping characters.  I
added the busy check and structured it a bit more like the Linux 1.0
kernel.  This also means that write() now correctly sets errno.

Also added a scheduling timeout when writing large buffers, and
cleaned up some port probing.