[bioshd] Implement park for ancient MFM/RLL disks
Really old drives, such as the two ST-412s in my IBM 5161, required
manual parking to ensure the r/w heads are over a safe location when the
disk spins down and the heads (intentionally!) crash onto the disk
surface.

In DOS land, this vital task was left up to the user, in the form of a
PARK command to be run before shutting down. In ELKS however, the kernel
should manage this automatically. An INT 13/AH=0Ch is invoked to seek
the disk to the last track when unmounted, and to all drives right
before halting the system.