[console] Add dual-screen support to console-direct.c
Thanks to some clever design from IBM folks, even the original 1981 IBM
PC had support for more than one display at a time by 1983. The CRTC I/O
base and vram segments don't always overlap, so nothing prevents the system
from addressing two compatible adapters simultaneously. The switch on
the PC motherboard just tells the BIOS which output it should set up on
boot.

This patch introduces the concept of an output to the console drivers,
and adds support for probing, setting up, and managing multiple
display adapters to the console-direct driver.

There are no new options. The adapter indicated by the BIOS as the
primary adapter is set up as the primary display, and a potential
secondary adapter will be set up as a secondary display, if present.
Depending on the runlevel, each display can get 0 or more consoles
allocated to it, and switching between them works the same as before,
by using CTRL+ALT+F<N>.
To indicate which display is active, the blinking cursor is toggled such
that it only appears on the active display.