romprg for the NEC V25 port
Minor changes to the 8018x romprg code to get it working on my NEC V25 system.  Just a "quick and dirty" job, as there is only a smal audience for this tool. Added a "how to" for my own reference. may be helpfull for others to, so I added it.